COMP: Fix conversion error in vtkPolyDataToitkMesh.cxx
This commit fixes the following build error:

/tmp/SPHARM-PDM/Modules/CLI/GenParaMeshCLP/vtkPolyDataToitkMesh.cxx:61:49: error: cannot convert ‘double*’ to ‘float*’ in initialization
         float * apoint = vtkpoints->GetPoint( p );
                                                 ^

+++ Modules/CLI/GenParaMeshCLP/vtkPolyDataToitkMesh.cxx
@@ -58,7 +58,7 @@ vtkPolyDataToitkMesh
     {
 
       #if VTK_MAJOR_VERSION > 5
-        float * apoint = vtkpoints->GetPoint( p );
+        double * apoint = vtkpoints->GetPoint( p );
       #else
         vtkFloatingPointType* apoint = vtkpoints->GetPoint( p );
       #endif
